+++ rpc/core/tx.go
@@ -3,11 +3,12 @@ package core
 import (
 	"fmt"
 
+	cmn "github.com/tendermint/tmlibs/common"
+
+	tmquery "github.com/tendermint/tendermint/libs/pubsub/query"
 	ctypes "github.com/tendermint/tendermint/rpc/core/types"
 	"github.com/tendermint/tendermint/state/txindex/null"
 	"github.com/tendermint/tendermint/types"
-	cmn "github.com/tendermint/tmlibs/common"
-	tmquery "github.com/tendermint/tmlibs/pubsub/query"
 )
 
 // Tx allows you to query the transaction results. `nil` could mean the