+++ state/txindex/null/null.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/tendermint/tendermint/state/txindex"
 	"github.com/tendermint/tendermint/types"
-	"github.com/tendermint/tmlibs/pubsub/query"
+	"github.com/tendermint/tendermint/libs/pubsub/query"
 )
 
 var _ txindex.TxIndexer = (*TxIndex)(nil)