+++ libs/pubsub/query/query_test.go
@@ -1,6 +1,7 @@
 package query_test
 
 import (
+	"fmt"
 	"testing"
 	"time"
 
@@ -12,38 +13,37 @@ import (
 )
 
 func TestMatches(t *testing.T) {
-	const shortForm = "2006-Jan-02"
-	txDate, err := time.Parse(shortForm, "2017-Jan-01")
-	require.NoError(t, err)
-	txTime, err := time.Parse(time.RFC3339, "2018-05-03T14:45:00Z")
-	require.NoError(t, err)
+	var (
+		txDate = "2017-01-01"
+		txTime = "2018-05-03T14:45:00Z"
+	)
 
 	testCases := []struct {
 		s       string
-		tags    map[string]interface{}
+		tags    map[string]string
 		err     bool
 		matches bool
 	}{
-		{"tm.events.type='NewBlock'", map[string]interface{}{"tm.events.type": "NewBlock"}, false, true},
+		{"tm.events.type='NewBlock'", map[string]string{"tm.events.type": "NewBlock"}, false, true},
 
-		{"tx.gas > 7", map[string]interface{}{"tx.gas": 8}, false, true},
-		{"tx.gas > 7 AND tx.gas < 9", map[string]interface{}{"tx.gas": 8}, false, true},
-		{"body.weight >= 3.5", map[string]interface{}{"body.weight": 3.5}, false, true},
-		{"account.balance < 1000.0", map[string]interface{}{"account.balance": 900}, false, true},
-		{"apples.kg <= 4", map[string]interface{}{"apples.kg": 4.0}, false, true},
-		{"body.weight >= 4.5", map[string]interface{}{"body.weight": float32(4.5)}, false, true},
-		{"oranges.kg < 4 AND watermellons.kg > 10", map[string]interface{}{"oranges.kg": 3, "watermellons.kg": 12}, false, true},
-		{"peaches.kg < 4", map[string]interface{}{"peaches.kg": 5}, false, false},
+		{"tx.gas > 7", map[string]string{"tx.gas": "8"}, false, true},
+		{"tx.gas > 7 AND tx.gas < 9", map[string]string{"tx.gas": "8"}, false, true},
+		{"body.weight >= 3.5", map[string]string{"body.weight": "3.5"}, false, true},
+		{"account.balance < 1000.0", map[string]string{"account.balance": "900"}, false, true},
+		{"apples.kg <= 4", map[string]string{"apples.kg": "4.0"}, false, true},
+		{"body.weight >= 4.5", map[string]string{"body.weight": fmt.Sprintf("%v", float32(4.5))}, false, true},
+		{"oranges.kg < 4 AND watermellons.kg > 10", map[string]string{"oranges.kg": "3", "watermellons.kg": "12"}, false, true},
+		{"peaches.kg < 4", map[string]string{"peaches.kg": "5"}, false, false},
 
-		{"tx.date > DATE 2017-01-01", map[string]interface{}{"tx.date": time.Now()}, false, true},
-		{"tx.date = DATE 2017-01-01", map[string]interface{}{"tx.date": txDate}, false, true},
-		{"tx.date = DATE 2018-01-01", map[string]interface{}{"tx.date": txDate}, false, false},
+		{"tx.date > DATE 2017-01-01", map[string]string{"tx.date": time.Now().Format(query.DateLayout)}, false, true},
+		{"tx.date = DATE 2017-01-01", map[string]string{"tx.date": txDate}, false, true},
+		{"tx.date = DATE 2018-01-01", map[string]string{"tx.date": txDate}, false, false},
 
-		{"tx.time >= TIME 2013-05-03T14:45:00Z", map[string]interface{}{"tx.time": time.Now()}, false, true},
-		{"tx.time = TIME 2013-05-03T14:45:00Z", map[string]interface{}{"tx.time": txTime}, false, false},
+		{"tx.time >= TIME 2013-05-03T14:45:00Z", map[string]string{"tx.time": time.Now().Format(query.TimeLayout)}, false, true},
+		{"tx.time = TIME 2013-05-03T14:45:00Z", map[string]string{"tx.time": txTime}, false, false},
 
-		{"abci.owner.name CONTAINS 'Igor'", map[string]interface{}{"abci.owner.name": "Igor,Ivan"}, false, true},
-		{"abci.owner.name CONTAINS 'Igor'", map[string]interface{}{"abci.owner.name": "Pavel,Ivan"}, false, false},
+		{"abci.owner.name CONTAINS 'Igor'", map[string]string{"abci.owner.name": "Igor,Ivan"}, false, true},
+		{"abci.owner.name CONTAINS 'Igor'", map[string]string{"abci.owner.name": "Pavel,Ivan"}, false, false},
 	}
 
 	for _, tc := range testCases {